Player takes damage from shots and can die now

+++ COP 4331 - Space Shooter 3D/Assets/Scripts/GameController.cs
@@ -54,7 +54,7 @@ public class GameController : MonoBehaviour
         playerHealth += val;
     }
 
-    public float GetHealth()
+    public int GetHealth()
     {
         return playerHealth;
     }
